refactor(wallets): move widget-only helpers out of wallets-shared
Relocate the pieces whose only consumers are widget packages, each to the lowest
package that needs it: `detectInstallLink` and its private `isBrave` to
`@rango-dev/ui`, `detectMobileScreens` to `@rango-dev/common-core`, and the
derivation path table to `@rango-dev/wallets-blockchains` as
`DERIVATION_PATHS_BY_NAMESPACE` alongside the `DerivationPath` type.

The table keeps only path data; nothing read `mainBlockchain` or `title`, so
those fields and the six namespaces carrying nothing else are gone. wallets-shared
drops `@hub3js/namespaces`, and DerivationPath.tsx loses an unused type import.
