+++ common/core/src/device.ts
@@ -1,0 +1,5 @@
+export function detectMobileScreens(): boolean {
+  return /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+    navigator.userAgent
+  );
+}
