+++ common/core/src/index.ts
@@ -1,1 +1,2 @@
 export * from './dynamicImport.js';
+export * from './device.js';
