+++ wallets/blockchains/src/mod.ts
@@ -1,2 +1,3 @@
 export * from './networks.js';
 export * from './caip.js';
+export * from './derivationPaths.js';
