+++ wallets/shared/package.json
@@ -28,7 +28,6 @@
   },
   "dependencies": {
     "@hub3js/core": "^0.61.1",
-    "@hub3js/namespaces": "^0.2.5",
     "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "ethers": "^6.13.2",
