+++ widget/embedded/package.json
@@ -26,6 +26,7 @@
     "@hub3js/namespaces": "^0.2.5",
     "@lingui/core": "4.2.1",
     "@lingui/react": "4.2.1",
+    "@rango-dev/common-core": "^0.1.0",
     "@rango-dev/logging-core": "^0.14.0",
     "@rango-dev/provider-all": "^0.65.0",
     "@rango-dev/queue-manager-core": "^0.34.0",
@@ -33,6 +34,7 @@
     "@rango-dev/queue-manager-react": "^0.34.0",
     "@rango-dev/signer-solana": "^0.49.0",
     "@rango-dev/ui": "^0.65.0",
+    "@rango-dev/wallets-blockchains": "^0.1.0",
     "@rango-dev/wallets-core": "^0.61.0",
     "@rango-dev/wallets-react": "^0.48.0",
     "@rango-dev/wallets-shared": "^0.62.0",
