+++ widget/ui/src/index.ts
@@ -1,7 +1,8 @@
 export * from './components/index.js';
 export * from './containers/index.js';
 export * from './theme.js';
-export * from './types/index.js';
+export type * from './types/index.js';
 export * from './hooks/index.js';
 export * from './icons/index.js';
 export * from './constants/index.js';
+export * from './utils/index.js';
